+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Fixed
+- Deprecation error in SettingsController.
+
 ## 1.0.0 2018-03-02
 
 The initial release of the Blockonomicon plugin.